api/requirements.txt
Django>=4.2,<5.0
djangorestframework==3.15.2
django-cors-headers==4.7.0